// src/app/shared/models/task.ts
export interface CustomerDemandReceipt {
  receipt_id: string;
  task_id: string;
  create_user: string;
  customer: string;
  title: string;
  create_date: string;
  estimate_over: string;
  task_over: string;
  task_manager: string;
  status: string;
  type: string;
  note: string;
}